This is how we should do it with references... I think this is the way to go
The tickets are now in their own collection, which gives us easy entry points to them. It's easy to fetch by assigned_to, etc.

It does mean that when fetching a milestone, you need to fetch the tickets separately. This isn't necessarily a bad thing, and it can be done by the service.

Let's hope mirmo agrees. :x

naneau out.